- Fixed bug #12375: Concurrency issues in images generation in a cluster environnement   (Merged from trunk (4.1.0alpha1) rev. 21504)
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/stable/3.10@21515 a01eee8c-daba-0310-acae-fa49f3023285

+++ lib/ezimage/classes/ezimagemanager.php
@@ -125,6 +125,7 @@ class eZImageManager
 
         $ini =& eZINI::instance( 'image.ini' );
         $this->TemporaryImageDirPath = eZSys::cacheDirectory() . '/' . $ini->variable( 'FileSettings', 'TemporaryDir' );
+        $this->lockTimeout = $ini->hasVariable( 'ImageConverterSettings', 'LockTimeout' ) ? $ini->variable( 'ImageConverterSettings', 'LockTimeout' ) : 60;
     }
 
     /*!
@@ -884,7 +885,6 @@ class eZImageManager
             $aliasKey = $currentAliasInfo['alias_key'];
 
             // VS-DBFILE
-
             require_once( 'kernel/classes/ezclusterfilehandler.php' );
             $aliasFile = eZClusterFileHandler::instance( $aliasFilePath );
 
@@ -899,17 +899,38 @@ class eZImageManager
                     $sourceMimeData['basename'] = $parameters['basename'];
                     eZMimeType::changeBasename( $destinationMimeData, $parameters['basename'] );
                 }
-                $destinationMimeData['is_valid'] = false;
-                if ( !$this->convert( $sourceMimeData, $destinationMimeData, $aliasName, $parameters ) )
+
+                if ( !$this->_exclusiveLock( $aliasFilePath, $aliasName ) )
                 {
-                    $sourceFile = $sourceMimeData['url'];
-                    $destinationDir = $destinationMimeData['dirpath'];
-                    eZDebug::writeError( "Failed converting $sourceFile to alias '$referenceAlias' in directory '$destinationDir'",
-                                         'eZImageManager::createImageAlias' );
-                    // VS-DBFILE
-                    $aliasFile->deleteLocal();
                     return false;
                 }
+
+                $wantImage = $this->imageAliasInfo( $sourceMimeData, $aliasName );
+                $wantImagePath = $wantImage['url'];
+
+                $fileHandler = eZClusterFileHandler::instance( $wantImagePath );
+                $fileHandler->loadMetaData( true );
+
+                if ( $fileHandler->exists() and $this->isImageTimestampValid( $fileHandler->mtime() ) )
+                {
+                    $destinationMimeData = $wantImage;
+                    $wasLocked = true;
+                }
+                else
+                {
+                    $destinationMimeData['is_valid'] = false;
+                    if ( !$this->convert( $sourceMimeData, $destinationMimeData, $aliasName, $parameters ) )
+                    {
+                        $sourceFile = $sourceMimeData['url'];
+                        $destinationDir = $destinationMimeData['dirpath'];
+                        eZDebug::writeError( "Failed converting $sourceFile to alias '$aliasName' in directory '$destinationDir'",
+                                             'eZImageManager::createImageAlias' );
+                        // VS-DBFILE
+                        $aliasFile->deleteLocal();
+                        return false;
+                    }
+                }
+
                 $currentAliasData = array( 'url' => $destinationMimeData['url'],
                                            'dirpath' => $destinationMimeData['dirpath'],
                                            'filename' => $destinationMimeData['filename'],
@@ -949,11 +970,19 @@ class eZImageManager
                             $currentAliasData['width'] = $width;
                             $currentAliasData['height'] = $height;
                         }
+                        else
+                        {
+                            eZDebug::writeError("The size of the generated image " . $destinationMimeData['url'] . " could not be read by getimagesize()", 'eZImageManager::createImageAlias' );
+                        }
 
                         // VS-DBFILE
 
-                        $fileHandler = eZClusterFileHandler::instance( $aliasFilePath );
-                        $fileHandler->fileStore( $destinationMimeData['url'], 'image', true, $destinationMimeData['name'] );
+                        // The file is not written to the database if it was already written due to a lock situation
+                        if ( !isset( $wasLocked ) )
+                        {
+                            $fileHandler = eZClusterFileHandler::instance( $aliasFilePath );
+                            $fileHandler->fileStore( $destinationMimeData['url'], 'image', true, $destinationMimeData['name'] );
+                        }
                     }
                     else
                     {
@@ -965,6 +994,9 @@ class eZImageManager
                 $existingAliasList[$aliasName] = $currentAliasData;
                 // VS-DBFILE
                 $aliasFile->deleteLocal();
+
+                $this->_freeExclusiveLock( $aliasFilePath, $aliasName );
+
                 return true;
             }
         }
@@ -980,6 +1012,7 @@ class eZImageManager
     function analyzeImage( &$mimeData, $parameters = array() )
     {
         $file = $mimeData['url'];
+
         if ( !file_exists( $file ) )
             return false;
         $analyzer = eZImageAnalyzer::createForMIME( $mimeData );
@@ -1014,6 +1047,7 @@ class eZImageManager
         include_once( 'lib/ezutils/classes/ezmimetype.php' );
         if ( is_string( $sourceMimeData ) )
             $sourceMimeData = eZMimeType::findByFileContents( $sourceMimeData );
+
         $this->analyzeImage( $sourceMimeData );
         $currentMimeData = $sourceMimeData;
         $handlers =& $this->ImageHandlers;
@@ -1023,6 +1057,7 @@ class eZImageManager
             $destinationPath = $destinationMimeData;
             $destinationMimeData = eZMimeType::findByFileContents( $destinationPath );
         }
+
         $filters = array();
         $alias = false;
         if ( $aliasName )
@@ -1270,6 +1305,42 @@ class eZImageManager
         return $result;
     }
 
+    /**
+     * Image information for $aliasName. This is the information which normally
+     * would be provided during generation of aliasName. This so that requests
+     * not holding the lock will provide meaningful information.
+     *
+     * @param mixed $mimeData 
+     * @param string $aliasName 
+     * @return mixed
+     */
+    function imageAliasInfo( $mimeData, $aliasName )
+    {
+        if ( is_string( $mimeData ) )
+            $mimeData = eZMimeType::findByFileContents( $mimeData );
+
+        $this->analyzeImage( $mimeData );
+        if ( $aliasName )
+        {
+            $aliasList = $this->aliasList();
+            if ( isset( $aliasList[$aliasName] ) )
+            {
+                $alias = $aliasList[$aliasName];
+                if ( $alias['mime_type'] )
+                {
+                    eZMimeType::changeMIMEType( $mimeData, $alias['mime_type'] );
+                }
+            }
+        }
+        if ( $aliasName != 'original' )
+        {
+            $mimeData['filename'] = $mimeData['basename'] . '_' . $aliasName . '.' . $mimeData['suffix'];
+            $mimeData['url'] = $mimeData['dirpath'] . '/' . $mimeData['filename'];
+        }
+        // eZDebug::writeDebug( $mimeData, 'return from eZImageManager::imageAliasInfo' );
+        return $mimeData;
+    }
+
     /*!
      \return the path for temporary images.
      \note The default value uses the temporary directory setting from site.ini.
@@ -1292,6 +1363,74 @@ class eZImageManager
         return $instance;
     }
 
+    /*!
+     Acquires an exclusive lock for the currently generated alias
+    */
+    function _exclusiveLock( $fileName, $aliasName )
+    {
+        // mutex w/ exclusive lock: convert(targetFileName)
+        $mutex =& $this->_mutex( "{$fileName}-{$aliasName}" );
+        while( true )
+        {
+            $timestamp  = $mutex->lockTS(); // Note: This does not lock, only checks what the timestamp is.
+            if ( $timestamp === false )
+            {
+                if ( !$mutex->lock() )
+                {
+                    eZDebug::writeWarning( "Failed to acquire lock for file $fileName/$aliasName" );
+                    return false;
+                }
+                $mutex->setMeta( 'pid', getmypid() );
+                return true;
+            }
+            if ( $timestamp >= gmmktime() - $this->lockTimeout )
+            {
+                sleep( 1 ); // Sleep 1 second
+                continue;
+            }
+
+            $oldPid = $mutex->meta( 'pid' );
+            if ( is_numeric( $oldPid ) &&
+                 $oldPid != 0 &&
+                 function_exists( 'posix_kill' ) )
+            {
+                posix_kill( $oldPid, 9 );
+            }
+            if ( !$mutex->steal() )
+            {
+                eZDebug::writeWarning( "Failed to steal lock for file $fileName/$aliasName from PID $oldPid" );
+                return false;
+            }
+            $mutex->setMeta( 'pid', getmypid() );
+            return true;
+        } // while
+    }
+
+    /*!
+     Frees the current exclusive lock in use.
+
+     \param $fname Name of the calling code (usually function name).
+     */
+    function _freeExclusiveLock( $fileName, $aliasName )
+    {
+        $mutex =& $this->_mutex( "{$fileName}-{$aliasName}" );
+        $mutex->unlock();
+    }
+
+    /*!
+     Returns the mutex object for the current file.
+     */
+    function &_mutex( $fname = false )
+    {
+        if ( $this->Mutex !== null )
+        {
+            return $this->Mutex;
+        }
+        include_once( 'lib/ezutils/classes/ezmutex.php' );
+        $this->Mutex = new eZMutex( $fname );
+        return $this->Mutex;
+    }
+
     /// \privatesection
     var $ImageHandlers;
     var $OutputMIME;
@@ -1301,6 +1440,22 @@ class eZImageManager
     var $RuleMap;
     var $MIMETypes;
     var $Types = array();
+
+    /**
+     * Holds the mutex for image alias file.
+     *
+     * @var eZMutex
+     */
+    var $Mutex;
+    
+    /**
+     * The time spent waiting before an existing eZMutex lock is cancelled and reused.
+     * Default value is 60 seconds, which is set in constructor.
+     *
+     * @var int
+     */
+    var $lockTimeout;
+    
 }
 
 ?>
